Add files via upload
Insertion sort.
Line 10: Usually the code of insertion sort does not include line 10 (when key of unsorted subarray is lesser than the last element of sorted subarray (j), placement of key is at initial position of j and then j is decreased). Usually, the placement of key is done at arr[j+1] but I could not decode the logic of this placement when key<arr[j]. It makes sense for key>arr[j] but not otherwise.